pyproject.toml
[project]
name = "website"
version = "0.1.0"
description = "Postfun backend and docs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Flask>=3.0.0",
  "Flask-Cors>=4.0.0",
  "Flask-SeaSurf>=1.1.1",
  "Flask-Limiter>=3.5.0",
  "Flask-Caching>=2.1.0",
  "redis>=5.0.0",
  "APScheduler>=3.10.4",
  "SQLAlchemy>=2.0.29",
  "Flask-SQLAlchemy>=3.1.1",
  "PyMySQL>=1.1.0",
  "PyJWT>=2.8.0",
  "coincurve>=18.0.0",
  "pynostr>=0.4.4",
  "bech32>=1.2.0",
  "python-dotenv>=1.0.0",
  "gunicorn>=21.2.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
